backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
psycopg2-binary
channels
channels-redis
Pillow
daphne
dj-database-url
python-dotenv
